refactor SwaggerGenOptionsExtensions
